requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cloudpickle==2.2.0
cryptography==38.0.1
dask==2022.10.0
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
fsspec==2022.10.0
gunicorn==20.1.0
importlib-metadata==5.0.0
locket==1.0.0
Markdown==3.4.1
packaging==21.3
partd==1.3.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
toolz==0.12.0
zipp==3.10.0
